requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-autocomplete-light==3.3.2
django-pwa==1.0.4
django-filter==2.1.0
django-rest-framework==0.1.0
djangorestframework==3.9.1
idna==2.8
oauthlib==3.0.1
psycopg2-binary==2.7.7
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
urllib3==1.24.1
coverage==4.5.3
django-nose==1.4.6
